feat: Logout && Delete User combined feature

+++ Blindar/View/Settings/SettingPage.swift
@@ -17,11 +17,11 @@ struct SettingPage: View {
     @State var isDailyNotificationOn: Bool = false
     @State var settingFeature: SettingFeature = .none
     @Binding var mainPageMode: MainPageMode
-    @State var showLogoutAlert: Bool = false
-    @State var showUserDeleteAlert: Bool = false
+    @State private var isLogoutAlertOn = false
+    @State private var isDeleteAlertOn = false
     
     var body: some View {
-        NavigationStack {
+        NavigationView {
             VStack {
                 Text("설정")
                     .font(.title)
@@ -62,45 +62,59 @@ struct SettingPage: View {
                         .foregroundColor(.white)
                     })
                 }
-                HStack(spacing: 70) {
-                    //로그아웃
+                .padding(.bottom, 30)
+                HStack {
+                    //로그아웃 및 회원탈퇴
                     Button(action: {
-                        showLogoutAlert = true
+                        isLogoutAlertOn = true
                     }, label: {
                         RoundedRectangle(cornerRadius: 8)
                             .foregroundColor(.hex00497B)
-                            .frame(width: uiManager.screenHeight * 0.11, height: uiManager.screenWidth * 0.1)
+                            .frame(height: 50)
                             .overlay {
-                                Text("로그아웃")
+                                Text("로그아웃 및 회원탈퇴")
                                     .foregroundStyle(Color.white)
                             }
                     })
-                    //회원 탈퇴
-                    Button(action: {
-                        showUserDeleteAlert = true
-                    }, label: {
-                        RoundedRectangle(cornerRadius: 8)
-                            .foregroundColor(.hex00497B)
-                            .frame(width: uiManager.screenHeight * 0.11, height: uiManager.screenWidth * 0.1)
-                            .overlay {
-                                Text("회원 탈퇴")
-                                    .foregroundStyle(Color.white)
-                            }
-                    })
+                    // 로그아웃 및 회원탈퇴 알림창
+                    .alert(isPresented: $isLogoutAlertOn) {
+                        Alert(
+                            title: Text("로그아웃 및 회원 탈퇴"),
+                            message: Text("회원과 관련된 모든 정보가 삭제됩니다"),
+                            primaryButton: .destructive(Text("확인"), action: {
+                                deleteUser()
+                            }),
+                            secondaryButton: .cancel(Text("취소"))
+                        )
+                    }
+                    //                    //회원 탈퇴
+                    //                    Button(action: {
+                    //                        isDeleteAlertOn = true
+                    //                    }, label: {
+                    //                        RoundedRectangle(cornerRadius: 8)
+                    //                            .foregroundColor(.hex00497B)
+                    //                            .frame(height: 50)
+                    //                            .overlay {
+                    //                                Text("회원 탈퇴")
+                    //                                    .foregroundStyle(Color.white)
+                    //                            }
+                    //                    })
+                    //                    // 회원 탈퇴 알림창
+                    //                    .alert(isPresented: $isDeleteAlertOn) {
+                    //                        Alert(
+                    //                            title: Text("회원 탈퇴"),
+                    //                            message: Text("회원과 관련된 모든 정보가 삭제됩니다"),
+                    //                            primaryButton: .destructive(Text("확인"), action: {
+                    //                                deleteUser()
+                    //                            }),
+                    //                            secondaryButton: .cancel(Text("취소"))
+                    //                        )
+                    //                    }
                 }
                 Spacer()
             }
             .padding(.horizontal, 12)
-        }
-        .alert(isPresented: $showLogoutAlert) {
-            Alert(title: Text("로그아웃"), primaryButton: .destructive(Text("확인"), action: {
-                logout()
-            }), secondaryButton: .cancel(Text("취소")))
-        }
-        .alert(isPresented: $showUserDeleteAlert) {
-            Alert(title: Text("회원 탈퇴"), message: Text("회원의 모든 정보가 삭제됩니다"), primaryButton: .destructive(Text("확인"), action: {
-                deleteUser()
-            }), secondaryButton: .cancel(Text("취소")))
+            .padding(.bottom)
         }
         .onAppear {
             isOnedayModeOn = mainPageMode == .oneday
@@ -108,37 +122,62 @@ struct SettingPage: View {
     }
     
     private func logout() {
+        print("로그아웃 함수 호출")
         let firebaseAuth = Auth.auth()
         do {
-          try firebaseAuth.signOut()
+            try firebaseAuth.signOut()
         } catch let signOutError as NSError {
-          print("Error signing out: %@", signOutError)
+            print("Error signing out: %@", signOutError)
         }
+        
+        //화면 이동
+        userVM.userState = .isNotRegistered
     }
     
     private func deleteUser() {
-        //firebase 계정 삭제
-        let user = Auth.auth().currentUser
+            if let user = Auth.auth().currentUser {
+                let userId = user.uid
+                let userName = userVM.user?.name ?? ""
 
-        user?.delete { error in
-          if let error = error {
-            print("An error happened")
-          } else {
-            print("Account deleted")
-          }
-        }
-        //로컬 데이터 삭제
-        do {
-            try modelContext.delete(model: MealLocalData.self)
-            try modelContext.delete(model: MemoLocalData.self)
-            try modelContext.delete(model: ScheduleLocalData.self)
-            try modelContext.delete(model: SchoolLocalData.self)
-            try modelContext.delete(model: UserLocalData.self)
-        } catch {
-            print("Failed to clear all Model Context data.")
-        }
-    }
-}
+                // Step 1: Delete user data from Realtime Database
+                userVM.tryDeleteUserFromFirebase(user: User(userId: userId, schoolCode: 0, name: userName, schoolName: ""))
+                    .sink(receiveCompletion: { completion in
+                        switch completion {
+                        case .failure(let error):
+                            print("Failed to delete user data from Firebase Database: \(error)")
+                        case .finished:
+                            // Step 2: Delete Firebase account
+                            user.delete { error in
+                                if let error = error {
+                                    print("Firebase Error: ", error)
+                                } else {
+                                    // Step 3: Clear local data
+                                    UserDefaults.standard.removeObject(forKey: "user")
+                                    
+                                    do {
+                                        try modelContext.delete(model: MealLocalData.self)
+                                        try modelContext.delete(model: MemoLocalData.self)
+                                        try modelContext.delete(model: ScheduleLocalData.self)
+                                        try modelContext.delete(model: SchoolLocalData.self)
+                                        try modelContext.delete(model: UserLocalData.self)
+                                        print("Local data deleted")
+                                    } catch {
+                                        print("Failed to clear all Model Context data: \(error.localizedDescription)")
+                                    }
+                                    
+                                    // Step 4: Logout
+                                    self.logout()
+                                }
+                            }
+                        }
+                    }, receiveValue: {
+                        
+                    })
+                    .store(in: &userVM.cancellables)
+            } else {
+                print("No login information available")
+            }
+        }}
 
 enum SettingFeature {
     case onedayMode
